User List Setting.
User List Setting.

+++ controller/AuthController.js
@@ -168,6 +168,24 @@ const signUpUserBySocialLogin = async (req, res) => {
     // Update user verification status to true
     user.gmailVerified = payload.email_verified;
     await user.save();
+
+    const userList = await UserListSchema.findOne({
+      userUuid: user.uuid
+    })
+
+    if(!userList){
+      const createUserList = new UserListSchema({
+        userUuid: user.uuid,
+      });
+      const newUserList = await createUserList.save();
+      if (!newUserList) {
+        await user.deleteOne({
+          uuid: uuid,
+        });
+        throw new Error("User not created due to list");
+      }
+    }
+
     // Create Ledger
     await createLedger({
       uuid: uuid,
@@ -304,15 +322,21 @@ const signUpUserBySocialBadges = async (req, res) => {
     // Update user verification status to true
     await user.save();
 
-    const createUserList = new UserListSchema({
-      userUuid: user.uuid,
-    });
-    const newUserList = await createUserList.save();
-    if (!newUserList) {
-      await user.deleteOne({
-        uuid: uuid,
+    const userList = await UserListSchema.findOne({
+      userUuid: user.uuid
+    })
+
+    if(!userList){
+      const createUserList = new UserListSchema({
+        userUuid: user.uuid,
       });
-      throw new Error("User not created due to list");
+      const newUserList = await createUserList.save();
+      if (!newUserList) {
+        await user.deleteOne({
+          uuid: uuid,
+        });
+        throw new Error("User not created due to list");
+      }
     }
 
     // Create Ledger
@@ -461,15 +485,21 @@ const createGuestMode = async (req, res) => {
     const users = await user.save();
     if (!users) throw new Error("User not Created");
 
-    const createUserList = new UserListSchema({
-      userUuid: users.uuid,
-    });
-    const newUserList = await createUserList.save();
-    if (!newUserList) {
-      await users.deleteOne({
-        uuid: uuid,
+    const userList = await UserListSchema.findOne({
+      userUuid: user.uuid
+    })
+
+    if(!userList){
+      const createUserList = new UserListSchema({
+        userUuid: user.uuid,
       });
-      throw new Error("User not created due to list");
+      const newUserList = await createUserList.save();
+      if (!newUserList) {
+        await user.deleteOne({
+          uuid: uuid,
+        });
+        throw new Error("User not created due to list");
+      }
     }
 
     // Generate a JWT token
@@ -518,7 +548,7 @@ const signUpGuestMode = async (req, res) => {
 
     const salt = await bcrypt.genSalt(10);
     const hashPassword = await bcrypt.hash(req.body.password, salt);
-    await User.updateOne(
+    const user = await User.updateOne(
       { uuid: req.body.uuid },
       {
         $set: {
@@ -529,6 +559,23 @@ const signUpGuestMode = async (req, res) => {
         },
       }
     );
+
+    const userList = await UserListSchema.findOne({
+      userUuid: user.uuid
+    })
+
+    if(!userList){
+      const createUserList = new UserListSchema({
+        userUuid: user.uuid,
+      });
+      const newUserList = await createUserList.save();
+      if (!newUserList) {
+        await user.deleteOne({
+          uuid: uuid,
+        });
+        throw new Error("User not created due to list");
+      }
+    }
 
     // Generate a JWT token
     const token = createToken({ uuid: req.body.uuid });
@@ -595,6 +642,24 @@ const signUpSocialGuestMode = async (req, res) => {
     // Update user verification status to true
     user.gmailVerified = payload.email_verified;
     await user.save();
+
+    const userList = await UserListSchema.findOne({
+      userUuid: user.uuid
+    })
+
+    if(!userList){
+      const createUserList = new UserListSchema({
+        userUuid: user.uuid,
+      });
+      const newUserList = await createUserList.save();
+      if (!newUserList) {
+        await user.deleteOne({
+          uuid: uuid,
+        });
+        throw new Error("User not created due to list");
+      }
+    }
+
     // Create Ledger
     await createLedger({
       uuid: uuid,
@@ -745,6 +810,23 @@ const signUpGuestBySocialBadges = async (req, res) => {
 
     // Update user verification status to true
     await user.save();
+
+    const userList = await UserListSchema.findOne({
+      userUuid: user.uuid
+    })
+
+    if(!userList){
+      const createUserList = new UserListSchema({
+        userUuid: user.uuid,
+      });
+      const newUserList = await createUserList.save();
+      if (!newUserList) {
+        await user.deleteOne({
+          uuid: uuid,
+        });
+        throw new Error("User not created due to list");
+      }
+    }
     // Create Ledger
     await createLedger({
       uuid: uuid,
